renaming demos
