+++ package.json
@@ -6,8 +6,13 @@
         "start": "node dist/app.js",
         "dev": "node --import ./register.js app.ts",
         "watch": "tsc -w",
-        "demo1": "node --import ./register.js demo-1.ts",
-        "demo2": "node --import ./register.js demo-2.ts",
+        "demo:helloworld": "node --import ./register.js demo-0-helloworld.ts",
+        "demo:city": "node --import ./register.js demo-city.ts",
+        "demo:prompt": "node --import ./register.js demo-1-prompt-configuration.ts",
+        "demo:structured": "node --import ./register.js demo-2-structured-output.ts",
+        "final:demo0": "node --import ./register.js final-demo-0.ts",
+        "final:demo1": "node --import ./register.js final-demo-1.ts",
+        "final:demo2": "node --import ./register.js final-demo-2-structured-output.ts",
         "image": "node --import ./register.js app-image.ts",
         "pdf": "node --import ./register.js app-pdf.ts",
         "board": "node --import ./register.js app-image-board.ts",
